+++ src/pages/NotFound.js
@@ -1,5 +1,7 @@
 import React from 'react';
 
 export default function NotFound() {
-  return <h2>404 - Not Found</h2>;
+  return (
+    <h2>404 - Not Found</h2>
+  );
 }
