+++ src/pages/monster-restaurant-docs/MRDocsHead.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+export function MRDocsHead(){
+
+    return (
+        <></>
+    );
+}
