+++ src/pages/portfolio/threejs/StaticValues.js
